Rename module node_registry -> registry
Renames the `splinter::node_registry` module to `splinter::registry`.
This is part of the effort to rebrand the "node registry" concept as
simply "registry".
