+++ libsplinter/src/lib.rs
@@ -74,11 +74,11 @@ pub mod keys;
 mod matrix;
 pub mod mesh;
 pub mod network;
-#[cfg(feature = "registry")]
-pub mod node_registry;
 pub mod orchestrator;
 pub mod protocol;
 pub mod protos;
+#[cfg(feature = "registry")]
+pub mod registry;
 #[cfg(feature = "rest-api")]
 pub mod rest_api;
 pub mod service;
