+++ libsplinter/src/registry/error.rs
