+++ libsplinter/src/registry/mod.rs
