+++ libsplinter/src/registry/rest_api/resources/nodes.rs
@@ -14,7 +14,7 @@
 
 use std::collections::HashMap;
 
-use crate::node_registry::Node;
+use crate::registry::Node;
 use crate::rest_api::paging::Paging;
 
 #[derive(Debug, Clone, PartialEq, Serialize)]
